初实现

+++ PhotonMapping/Material/color.cpp
@@ -11,11 +11,11 @@ Color operator * (const Color& A, const Color& B) {
 	return Color(A.r * B.r, A.g * B.g, A.b * B.b);
 }
 
-Color operator * (const Color& A, float k) {
+Color operator * (const Color& A, double k) {
 	return Color(A.r * k, A.g * k, A.b * k);
 }
 
-Color operator / (const Color& A, float k) {
+Color operator / (const Color& A, double k) {
 	return Color(A.r / k, A.g / k, A.b / k);
 }
 
@@ -29,21 +29,21 @@ Color& operator -= (Color& A, const Color& B) {
 	return A;
 }
 
-Color& operator *= (Color& A, float k) {
+Color& operator *= (Color& A, double k) {
 	A = A * k;
 	return A;
 }
 
-Color& operator /= (Color& A, float k) {
+Color& operator /= (Color& A, double k) {
 	A = A / k;
 	return A;
 }
 
 Color Color::Confine() {
-	return Color(std::min(r, 1.0f), std::min(g, 1.0f), std::min(b, 1.0f));
+	return Color(std::min(r, 1.0), std::min(g, 1.0), std::min(b, 1.0));
 }
 
-float Color::Power() {
+double Color::Power() {
 	return (r + g + b) / 3;
 }
