+++ PhotonMapping/Math/vector3.cpp
@@ -8,11 +8,11 @@ Vector3 operator - (const Vector3& A, const Vector3& B) {
 	return Vector3(A.x - B.x, A.y - B.y, A.z - B.z);
 }
 
-Vector3 operator * (const Vector3& A, float k) {
+Vector3 operator * (const Vector3& A, double k) {
 	return Vector3(A.x * k, A.y * k, A.z * k);
 }
 
-Vector3 operator / (const Vector3& A, float k) {
+Vector3 operator / (const Vector3& A, double k) {
 	return Vector3(A.x / k, A.y / k, A.z / k);
 }
 
@@ -30,12 +30,12 @@ Vector3& operator -= (Vector3& A, const Vector3& B) {
 	return A;
 }
 
-Vector3& operator *= (Vector3& A, float k) {
+Vector3& operator *= (Vector3& A, double k) {
 	A = A * k;
 	return A;
 }
 
-Vector3& operator /= (Vector3& A, float k) {
+Vector3& operator /= (Vector3& A, double k) {
 	A = A / k;
 	return A;
 }
@@ -49,33 +49,33 @@ Vector3 operator - (const Vector3& A) {
 	return Vector3(-A.x, -A.y, -A.z);
 }
 
-float Vector3::Dot(const Vector3& term) {
+double Vector3::Dot(const Vector3& term) {
 	return x * term.x + y * term.y + z * term.z;
 }
 
-inline float Vector3::Module2() {
+inline double Vector3::Module2() {
 	return x * x + y * y + z * z;
 }
 
-inline float Vector3::Module(){
+inline double Vector3::Module(){
 	return sqrt(x * x + y * y + z * z);
 }
 
-float Vector3::Distance2(Vector3& term){
+double Vector3::Distance2(Vector3& term){
 	return (term - *this).Module2();
 }
 
-float Vector3::Distance(Vector3& term) {
+double Vector3::Distance(Vector3& term) {
 	return (term - *this).Module();
 }
 
-float Vector3::IncludedAngle(Vector3 v2)
+double Vector3::IncludedAngle(Vector3 v2)
 {
-	float costheta = (this->Dot(v2)) / (this->Module()*v2.Module());
+	double costheta = (this->Dot(v2)) / (this->Module()*v2.Module());
 	return acos(costheta);
 }
 
-float& Vector3::GetCoord(int axis){
+double& Vector3::GetCoord(int axis){
 	if (axis == 0) return x;
 	if (axis == 1) return y;
 	return z;
@@ -109,16 +109,16 @@ Vector3 Vector3::Reflect(Vector3 N) {
 	return *this - N * (2 * Dot(N));
 }
 
-Vector3 Vector3::Refract(Vector3 N, float n) {
+Vector3 Vector3::Refract(Vector3 N, double n) {
 	Vector3 V = GetUnitVector();
-	float cosI = -N.Dot(V), cosT2 = 1 - (n * n) * (1 - cosI * cosI);
+	double cosI = -N.Dot(V), cosT2 = 1 - (n * n) * (1 - cosI * cosI);
 	if (cosT2 > EPS) return V * n + N * (n * cosI - sqrt(cosT2));
 	return V.Reflect(N);
 }
 
 Vector3 Vector3::Diffuse(Vector3 N) {
-	float theta = RandomRealZeroOne() *2*PI;
-	float random2 = RandomRealZeroOne();
+	double theta = RandomRealZeroOne() *2*PI;
+	double random2 = RandomRealZeroOne();
 	Vector3 z = N.GetUnitVector();
 	Vector3 y;
 	/*大于0.1 与y轴作叉积,小于0.1与x轴作叉积*/
