+++ PhotonMapping/Scene/camera.cpp
@@ -1,7 +1,7 @@
 #include "camera.h"
 
 
-Vector3 Camera::Emit(float i, float j) {
+Vector3 Camera::Emit(double i, double j) {
 	return lookAt + dh * (2 * i / H - 1) + dw * (2 * j / W - 1);
 }
 void Camera::Output(Bmp* bmp) {
