+++ PhotonMapping/include/ray.h
@@ -8,7 +8,7 @@ public:
 	Ray(Vector3 _origin = Vector3(0, 0, 0), Vector3 _direction = Vector3(0, 0, 0)) :origin(_origin), direction(_direction) {}
 	void rayDiff(Vector3 normal);
 	void rayRefl(Vector3 normal);
-	void rayRefr(Vector3 normal, float n);
+	void rayRefr(Vector3 normal, double n);
 	
 };
